backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
distlib==0.3.7
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
filelock==3.12.4
Markdown==3.5
pipenv==2023.10.20
platformdirs==3.11.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
virtualenv==20.24.5
djangorestframework-simplejwt